update: Create lib/utils.js
